requirements.txt
rich
pyyaml
SQLAlchemy
fastapi
uvicorn
httpx
aiosqlite
greenlet
websockets
jinja2
uro
